student grade

+++ internal/entity/grade_component_entity.go
@@ -2,6 +2,6 @@ package entity
 
 type GradeComponent struct {
 	Entity
-	Name   string `gorm:"column:name"`
-	Weight int    `gorm:"column:weight"`
+	Name   string  `gorm:"column:name"`
+	Weight float64 `gorm:"column:weight"`
 }
